# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /superadmin/schools:
    get:
      operationId: listSchools
      tags: [superadmin]
      summary: List all schools
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [superadmin]
      summary: Create a new school
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolRequest"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /superadmin/schools/{schoolId}:
    get:
      operationId: getSchoolById
      tags: [superadmin]
      summary: Get a school by ID
      security:
        - bearerAuth: []
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    put:
      operationId: updateSchool
      tags: [superadmin]
      summary: Update school details
      security:
        - bearerAuth: []
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchoolRequest"
      responses:
        "200":
          description: School updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    delete:
      operationId: deleteSchool
      tags: [superadmin]
      summary: Delete a school
      security:
        - bearerAuth: []
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ---- SCHOOL ADMIN ----
...
  /student/profile:
    get:
      operationId: getStudentProfile
      tags: [student]
      summary: Get student profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /student/results:
    get:
      operationId: getStudentResults
      tags: [student]
      summary: Get student results
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
  /student/notifications:
    get:
      operationId: getStudentNotifications
      tags: [student]
      summary: Get student notifications from their school
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /student/change-password:
    post:
      operationId: changeStudentPassword
      tags: [student]
      summary: Change student password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    School:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        username:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - username
        - isActive
        - createdAt
    CreateSchoolRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
      required:
        - name
        - slug
        - username
        - password
    UpdateSchoolRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        isActive:
          type: boolean
        password:
          type: string
          nullable: true
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - isActive
        - createdAt
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rollNumber:
          type: string
        className:
          type: string
          nullable: true
        section:
          type: string
          nullable: true
        fatherName:
          type: string
          nullable: true
        motherName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        enrollmentDate:
          type: string
          nullable: true
        schoolName:
          type: string
        schoolSlug:
          type: string
        hasChangedPassword:
          type: boolean
      required:
        - id
        - name
        - rollNumber
        - schoolName
        - schoolSlug
        - hasChangedPassword
...
    Result:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
          nullable: true
        rollNumber:
          type: string
          nullable: true
        className:
          type: string
          nullable: true
        subject:
          type: string
        marks:
          type: number
        maxMarks:
          type: number
        grade:
          type: string
          nullable: true
        examType:
          type: string
          nullable: true
        examDate:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
      required:
        - id
        - studentId
        - subject
        - marks
        - maxMarks
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword