# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfile {
  id: number;
  name: string;
  rollNumber: string;
  className?: string | null;
  section?: string | null;
  fatherName?: string | null;
  motherName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  enrollmentDate?: string | null;
  schoolName: string;
  schoolSlug: string;
  hasChangedPassword: boolean;
}