# lib/api-zod/src/generated/types/updateSchoolRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSchoolRequest {
  name?: string;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  isActive?: boolean;
  password?: string | null;
}